Moved maze functions, removed unused functions
